% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_to_local}
\alias{copy_to_local}
\title{Mirror Files from Network Drive to Local Project}
\usage{
copy_to_local(
  relative_paths,
  source_root = Sys.getenv("V_DRIVE"),
  local_root = "./data"
)
}
\arguments{
\item{relative_paths}{Character vector of paths relative to `source_root`.}

\item{source_root}{Base directory (defaults to `V_DRIVE` env var).}

\item{local_root}{Destination directory (defaults to `./data`).}
}
\description{
Efficiently copies files from a source root to a local directory using
`robocopy` (Windows) or `rsync` (macOS/Linux). \cr
Assumes destination folder name and V_DRIVE env. var.
}
