% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_utility}
\alias{copy_utility}
\title{Efficiently copies specific files or directories to a target directory.}
\usage{
copy_utility(relative_paths, destination_folder = "", source_root = "")
}
\arguments{
\item{relative_paths}{File or folder paths to copy relative to `source_root`.}

\item{destination_folder}{Exact folder path where files will be copied to.}

\item{source_root}{Base/project directory.}
}
\description{
Efficiently copies specific files or directories to a target directory.
}
