% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vd_fp}
\alias{vd_fp}
\title{Construct Paths Relative to user-environment-specific V-Drive Root}
\usage{
vd_fp(...)
}
\arguments{
\item{...}{Character vector passed to \code{file.path()}.}
}
\value{
A character vector of the full file path.
}
\description{
Prefixes paths with the "V_DRIVE" environment variable. \cr
Set it by running \code{usethis::edit_r_environ()} in the console, opening
.Renviron, and adding V_DRIVE = yourVDRIVEpath.\cr
Facilitates script collaboration by removing hard-coded VDRIVE which may
appear differently to different users.\cr
}
